services/parser/pyproject.toml
[tool.poetry]
name = "parser"
version = "0.1.0"
description = ""
authors = ["Serafim Oleynik <serafim.oleynik.5777@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pydantic = "^2.6.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.1.0"
fake-useragent = "^1.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[tool.poetry.scripts]
main = "parser.main:main"
scenarios = "parser.scenarios.full_bypass:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
